write trial key to gon in all places that we need it, not only in enterprise module
