+++ app/views/common/upsale.html.erb
@@ -1,5 +1,9 @@
 <% write_augur_to_gon %>
 
+<% if !EnterpriseToken.current.present? %>
+  <% write_trial_key_to_gon %>
+<% end %>
+
 <div class="upsale-notification">
     <h2><%= feature_title %></h2>
     <% if feature_description.empty? %>
